Cambiado telefono de instructor a double y todo los cambios extras necesarios, añadido InstructorValidador, modificado add.jsp de instructor

+++ ProyectoNaturAdventureWeb/src/es/uji/ei1027/naturAdventure/domain/Profile.java
@@ -17,8 +17,8 @@ public interface Profile {
 	public void setEmail( String email );
 	public String getEmail();
 	
-	public void setTelephone( int telephone );
-	public int getTelephone();
+	public void setTelephone( long telephone );
+	public long getTelephone();
 	
 	public void setUsername( String username );
 	public String getUsername();
